Implement template for Nicoles new UI
