+++ frontend/scripts/navigation.ts
@@ -1,45 +1,23 @@
+const mainMenu = document.getElementById("main-menu");
+const navBar = document.getElementById("nav-bar");
+const body = document.body;
 
-import { newPage } from './index.js'
-
-const nav = new Map<string, string>([
-    ["Home", "/"],
-    ["Game", "/game"],
-    ["Tournament", "/lobby"],
-    // ["Chat", "/chat"],
-    ["Login/Sign In","/login"],
-    ["My Home Page", "/mypage"],
-]);
-
-export function AddNavigation()
-{
-    const navBar = document.createElement("nav");
-
-    navBar.id = "nav";
-    navBar.className = "flex bg-gray-600 p-6";
-
-    nav.forEach((v, k, m) =>
-    {
-        const link = document.createElement("a");
-        link.text = k;
-        link.href = v;
-        link.className = "block mt-4 mr-4"
-        link.onclick = NavOnClick;
-        navBar.appendChild(link);
-        navBar.appendChild(document.createElement("br"));
-    })
-    document.body.appendChild(navBar);
+function openMenu() {
+    if (mainMenu) {
+        mainMenu.style.height = "100vh";
+        body.style.overflow = "";
+    }
+    else
+        console.error("ERROR: mainMenu element not found.");
 }
 
-export function NavOnClick(e: MouseEvent)
-{
-    if (!(e.target instanceof HTMLAnchorElement))
-        return;
-
-    e.preventDefault();
-
-    const newURL = e.target.href;
-
-    history.pushState({}, "", newURL);
-
-    newPage();
+function closeMenu() {
+    if (mainMenu) {
+        mainMenu.style.height = "0";
+        body.style.overflow = "hidden";
+    }
+    else
+        console.error("ERROR: mainMenu element not found.");
 }
+
+(window as any).openMenu = openMenu;
